ml/training/retrain_pipeline.py
"""
Retraining pipeline for the TerraFusion ML models.

This module provides infrastructure for retraining models when drift is detected
or on a scheduled basis. It handles the full model lifecycle from training to
evaluation to deployment.
"""

import os
import logging
import json
from datetime import datetime
import pandas as pd
import joblib
import numpy as np
from typing import Dict, Any, List, Tuple

from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger('retrain_pipeline')

class RetrainPipeline:
    """Pipeline for retraining and updating the levy impact prediction model."""

    def __init__(self, 
                 data_path: str = 'data/levy_training_data.csv',
                 model_path: str = 'ml/models/levy_impact_model.pkl',
                 metrics_path: str = 'ml/models/model_metrics.json'):
        """
        Initialize the retraining pipeline.
        
        Args:
            data_path: Path to the training data
            model_path: Path to save the trained model
            metrics_path: Path to save model metrics
        """
        self.data_path = data_path
        self.model_path = model_path
        self.metrics_path = metrics_path
        
        # Create directories if they don't exist
        os.makedirs(os.path.dirname(self.model_path), exist_ok=True)

    def load_data(self) -> Tuple[pd.DataFrame, pd.DataFrame, pd.Series, pd.Series]:
        """
        Load and prepare the data for training.
        
        Returns:
            Tuple of (X_train, X_test, y_train, y_test)
        """
        logger.info(f"Loading data from {self.data_path}")
        
        try:
            # Load data
            data = pd.read_csv(self.data_path)
            
            # Define features and target
            target_col = 'forecast'
            feature_cols = [col for col in data.columns if col != target_col]
            
            # Create features and target
            X = data[feature_cols]
            y = data[target_col]
            
            # Split into train/test (80/20 split)
            # In a real implementation, this would use a more sophisticated approach
            # like time-based split for time series data
            test_size = int(len(data) * 0.2)
            X_train = X.iloc[:-test_size]
            X_test = X.iloc[-test_size:]
            y_train = y.iloc[:-test_size]
            y_test = y.iloc[-test_size:]
            
            logger.info(f"Data loaded: {len(X_train)} training samples, {len(X_test)} test samples")
            
            return X_train, X_test, y_train, y_test
            
        except Exception as e:
            logger.error(f"Error loading data: {str(e)}")
            raise

    def train_model(self, X_train: pd.DataFrame, y_train: pd.Series) -> Any:
        """
        Train a new model on the provided data.
        
        Args:
            X_train: Training features
            y_train: Training target
            
        Returns:
            Trained model
        """
        logger.info("Training new model")
        
        try:
            # In a real implementation, this would include:
            # - Hyperparameter tuning
            # - Cross-validation
            # - Feature selection
            # - Multiple model comparison
            
            # For simplicity, we'll use a RandomForest regression model
            model = RandomForestRegressor(
                n_estimators=100,
                max_depth=10,
                random_state=42
            )
            
            # Train the model
            model.fit(X_train, y_train)
            
            logger.info("Model training completed successfully")
            
            return model
            
        except Exception as e:
            logger.error(f"Error training model: {str(e)}")
            raise

    def evaluate_model(self, model: Any, X_test: pd.DataFrame, y_test: pd.Series) -> Dict[str, float]:
        """
        Evaluate the trained model on the test data.
        
        Args:
            model: Trained model
            X_test: Test features
            y_test: Test target
            
        Returns:
            Dictionary of evaluation metrics
        """
        logger.info("Evaluating model performance")
        
        try:
            # Make predictions
            y_pred = model.predict(X_test)
            
            # Calculate metrics
            mse = mean_squared_error(y_test, y_pred)
            mae = mean_absolute_error(y_test, y_pred)
            r2 = r2_score(y_test, y_pred)
            
            # Log results
            logger.info(f"Model evaluation results: MSE={mse:.4f}, MAE={mae:.4f}, R²={r2:.4f}")
            
            # Return metrics
            return {
                'mse': mse,
                'mae': mae,
                'r2': r2,
                'timestamp': datetime.now().isoformat()
            }
            
        except Exception as e:
            logger.error(f"Error evaluating model: {str(e)}")
            raise

    def save_model(self, model: Any, metrics: Dict[str, float]) -> bool:
        """
        Save the trained model and metrics.
        
        Args:
            model: Trained model
            metrics: Evaluation metrics
            
        Returns:
            Boolean indicating success
        """
        logger.info(f"Saving model to {self.model_path}")
        
        try:
            # Save model
            joblib.dump(model, self.model_path)
            
            # Add model version and other metadata to metrics
            metrics['model_version'] = '1.0'
            metrics['training_size'] = model.n_estimators
            
            # Save metrics
            with open(self.metrics_path, 'w') as f:
                json.dump(metrics, f, indent=2)
            
            # Create training history file if it doesn't exist
            history_path = os.path.join(os.path.dirname(self.model_path), 'training_history.csv')
            if not os.path.exists(history_path):
                with open(history_path, 'w') as f:
                    f.write('timestamp,mse,mae,r2,model_version\n')
            
            # Append to training history
            with open(history_path, 'a') as f:
                f.write(f"{metrics['timestamp']},{metrics['mse']},{metrics['mae']},{metrics['r2']},{metrics['model_version']}\n")
            
            logger.info("Model and metrics saved successfully")
            
            return True
            
        except Exception as e:
            logger.error(f"Error saving model: {str(e)}")
            return False

    def run(self) -> Dict[str, float]:
        """
        Run the complete retraining pipeline.
        
        Returns:
            Dictionary of final evaluation metrics
        """
        logger.info("Starting retraining pipeline")
        
        try:
            # Load data
            X_train, X_test, y_train, y_test = self.load_data()
            
            # Train model
            model = self.train_model(X_train, y_train)
            
            # Evaluate model
            metrics = self.evaluate_model(model, X_test, y_test)
            
            # Save model and metrics
            self.save_model(model, metrics)
            
            logger.info("Retraining pipeline completed successfully")
            
            return metrics
            
        except Exception as e:
            logger.error(f"Error in retraining pipeline: {str(e)}")
            raise

if __name__ == "__main__":
    # Create and run pipeline
    pipeline = RetrainPipeline()
    try:
        metrics = pipeline.run()
        print(f"Retraining complete: MSE={metrics['mse']:.4f}, MAE={metrics['mae']:.4f}, R²={metrics['r2']:.4f}")
    except Exception as e:
        print(f"Error: {str(e)}")