pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "markupsafe>=3.0.2",
    "numpy>=2.2.4",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "statsmodels>=0.14.4",
    "werkzeug>=3.1.3",
    "xlrd>=2.0.1",
    "click>=8.1.8",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "trafilatura>=2.0.0",
    "polars-distance==0.5.1",
    "polars-business==0.3.22",
    "polars==1.16.0",
    "polars-xdt==0.16.0",
    "polarion==1.4.0",
    "polars-coord-transforms==0.10.0",
    "polars-ols==0.3.5",
    "polaris==0.1",
    "polartransform>=2.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "alembic>=1.15.2",
    "sqlalchemy-utils>=0.41.2",
    "prettytable>=3.16.0",
]
